requirements.txt
alembic==1.10.3
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
greenlet==2.0.2
importlib-metadata==6.2.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
